% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_info_assoc.R
\name{run_info_assoc}
\alias{run_info_assoc}
\title{Run enrichment analysis using information theory. Which pathways explain the phenotype}
\usage{
run_info_assoc(
  expr,
  gene_sets,
  phenotype,
  universe = NULL,
  score = c("mean_z", "pc1"),
  nbins = 10,
  n_perm = 1000,
  min_set_size = 10,
  seed = NULL
)
}
\arguments{
\item{expr}{samples x proteins/genes}

\item{gene_sets}{list of characters with relevant pathways.}

\item{phenotype}{outcome vector}

\item{universe}{Character vector of all genes in the background set.}

\item{n_perm}{Number of permutations for null model.}

\item{seed}{Optional random seed for reproducibility.}
}
\value{
A dataframe with results
}
\description{
Run enrichment analysis using information theory. Which pathways explain the phenotype
}
